Moved queries from allCountries.java to sql with the prints moved to Reports and removed unnecessary  variables from Country
