--- src/main/java/com/napier/sem/Country.java
@@ -1,20 +1,0 @@
-package com.napier.sem;
-
-public class Country {
-
-    public String code;
-    public String name;
-    public String continent;
-    public String region;
-    public float surfacearea;
-    public int indeptyear;
-    public int population;
-    public float lifeexpectancy;
-    public float gnp;
-    public float gnpold;
-    public String localname;
-    public String governmentform;
-    public String headofstate;
-    public int capital;
-    public String code2;
-}
