+++ src/main/java/com/napier/sem/report/Reports.java
@@ -1,6 +1,7 @@
 package com.napier.sem.report;
 
 import com.napier.sem.world.City;
+import com.napier.sem.world.Country;
 import com.napier.sem.world.PopulationData;
 
 import java.util.ArrayList;
@@ -70,6 +71,60 @@ public class Reports {
             String dataString = String.format("%-30s %15s %15s %15s",
                     populationData.identifier, populationData.population, populationData.populationInCities, populationData.populationNotInCities);
             System.out.println(dataString);
+        }
+    }
+
+    /**
+     * Method to print all countries and their population from largest to smallest
+     * @param countries
+     */
+    public void printCountriesInWorld(ArrayList<Country> countries)
+    {
+        // Print header
+        System.out.println(String.format("%-10s %-15s", "Name", "Population"));
+        // Loop over all employees in the list
+        for (Country country : countries)
+        {
+            String country_string =
+                    String.format("%-10s %-15s",
+                            country.name, country.population);
+            System.out.println(country_string);
+        }
+    }
+
+    /**
+     * Method to print countries and their population of a specified continent from smallest to largest
+     * @param countries
+     */
+    public void printCountriesInContinent(ArrayList<Country> countries)
+    {
+        // Print header
+        System.out.println(String.format("%-10s %-15s %-10s", "Name", "Population", "Continent"));
+        // Loop over all employees in the list
+        for (Country country : countries)
+        {
+            String country_string =
+                    String.format("%-10s %-15s %-10s",
+                            country.name, country.population, country.continent);
+            System.out.println(country_string);
+        }
+    }
+
+    /**
+     * Method to print countries and their population of a specified region from smalled to largest
+     * @param countries
+     */
+    public void printCountriesInRegion(ArrayList<Country> countries)
+    {
+        // Print header
+        System.out.println(String.format("%-10s %-15s %-10s", "Name", "Population", "Region"));
+        // Loop over all employees in the list
+        for (Country country : countries)
+        {
+            String country_string =
+                    String.format("%-10s %-15s %-10s",
+                            country.name, country.population, country.region);
+            System.out.println(country_string);
         }
     }
 }
