+++ src/main/java/com/napier/sem/world/Country.java
@@ -1,0 +1,10 @@
+package com.napier.sem.world;
+
+public class Country {
+
+
+    public String name;
+    public String continent;
+    public String region;
+    public int population;
+}
